requirements.txt
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
xgboost==3.2.0
shap==0.51.0
joblib==1.5.3
fastapi
uvicorn
streamlit
requests
pytesseract
pillow
groq